--- enumerable-method.rb
@@ -1,103 +1,0 @@
-module Enumerable
-    def my_each
-        i = 0
-       loop  do
-            if i < self.length
-          yield self[i] if block_given?
-          i += 1
-      else
-          break
-        end
-    end
-    self
-      end
-
-      def my_each_with_index
-        i = 0
-       loop  do
-            if i < self.length
-          yield self[i], i if block_given?
-          i += 1
-      else
-          break
-        end
-    end
-    self
-      end
-
-      def my_select
-        i = 0
-        newArr = []
-        loop do
-             if i < self.length
-         newArr << self[i] if yield self[i]
-           i += 1
-       else
-           break
-        end 
-    end
-    newArr
-     end
-
-     def my_all?
-        result = true
-         my_each do |num|
-         unless yield num
-          result = false
-         end
-       end
-       result
-      end
-
-      def my_any?
-        result = false
-       my_each do |num|
-        if yield num
-        result = true 
-    end
-     end
-     result
-    end
-
-    def my_none?
-        result = true
-       my_each do |num|
-        if yield num
-        result = false 
-      end
-      end
-      result
-      end
-
-      def my_count
-        count = 0
-        my_each do |num|
-            if yield num
-            count += 1
-        end
-        end
-        count
-       end
-
-       def my_map
-        arr = []
-      my_each do |num|
-        arr << yield(num) if yield(num)
-      end
-      arr
-      end
-
-      def my_inject(&proc)
-        sum = self[0]
-        self.drop(1).my_each do |num|
-          sum = proc.call(sum, num)
-        end
-        sum
-      end
-end
-
-def multiply_els(arr)
-    arr.my_inject { |x, y| x * y}
-  end
-  
-  p multiply_els([2,4,5])
